shadow-dom/styles/test-006.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_06_00_07</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#styles">
<meta name="assert" content="Styles:Rules that contain select reference combinators match elements in the enclosing trees">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
//Reference combinators are not implemented yet, so the test shouldn't work for now
//See the progress at  http://wkb.ug/82169
test(unit(function (ctx) {
	var d = newRenderedHTMLDocument(ctx);

    d.body.innerHTML =
    	'<ul class="cls">' +
    		'<li id="li1" class="shadow">1</li>' +
    		'<li id="li2" class="shadow2">2</li>' +
    		'<li id="li3" class="shadow">3</li>' +
    		'<li id="li4">4</li>' +
    		'<li id="li5" class="shadow">5</li>' +
    		'<li id="li6" class="shadow2">6</li>' +
    	'</ul>' +
    	'<div id="sr">' +
    	'</div>';

	var host = d.querySelector('#sr');

	//Shadow root to play with
	var s = createSR(host);

	var style = d.createElement('style');
	style.innerHTML ='ul.cls/select/li.shadow {display:none}';
	s.appendChild(style);

	var span = d.createElement('span');
	span.setAttribute('id', 'theShadowSpan');
	span.setAttribute('class', 'invis');
	s.appendChild(span);

	//li1, li3 and li5 should be invisible
	assert_equals(d.querySelector('#li1').offsetTop, 0,
		'Point 1: Rules that contain select reference combinators should match elements ' +
		'in the enclosing trees');
	assert_equals(d.querySelector('#li3').offsetTop, 0,
		'Point 2: Rules that contain select reference combinators should match elements ' +
		'in the enclosing trees');
	assert_equals(d.querySelector('#li5').offsetTop, 0,
		'Point 3: Rules that contain select reference combinators should match elements ' +
		'in the enclosing trees');

}), 'A_06_00_07_T01');
</script>
</body>
</html>
